+++ ex00/src/main/java/edu/school21/cinema/models/User.java
