+++ ex00/src/main/java/edu/school21/cinema/servlets/SignInServlet.java
@@ -1,0 +1,42 @@
+package edu.school21.cinema.servlets;
+
+import edu.school21.cinema.services.UserHandler;
+import org.springframework.context.ApplicationContext;
+
+import javax.servlet.ServletConfig;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import java.io.IOException;
+
+@WebServlet("/signIn")
+public class SignInServlet extends HttpServlet {
+
+    private ApplicationContext springContext;
+
+    @Override
+    public void init(ServletConfig config) {
+        springContext = (ApplicationContext) config.getServletContext().getAttribute("springContext");
+    }
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        req.getRequestDispatcher("/WEB-INF/html/signIn.html").forward(req, resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+
+        UserHandler userHandler = springContext.getBean("userHandler", UserHandler.class);
+
+        if (userHandler.read(req.getParameter("email"), req.getParameter("password"))){
+            HttpSession session = req.getSession();
+            session.setAttribute("user", userHandler.get(req.getParameter("email")));
+            req.getRequestDispatcher("/WEB-INF/html/welcome.html").forward(req, resp);
+        }
+        else doGet(req, resp);
+    }
+}
