+++ ex00/src/main/resources/sql/data.sql
@@ -1,0 +1,16 @@
+insert into users (name, surname, phone, email, password) VALUES
+    (
+        'Alex',
+        'Inkov',
+        '+79851961010',
+        'heigan@gmail.com',
+        '$2a$10$y2VLvkxGnUv5LnOTPDF8VuusJYURmEshgXAqaxe5BOb/oLUSjgsJK'
+    );
+insert into users (name, surname, phone, email, password) VALUES
+    (
+        'German',
+        'Gref',
+        '+75555555555',
+        'glavny@sberbank.ru',
+        '$2a$10$IiA//xazjcOy1yFL9p6SoujSUWZtKaTkS0/AQFGgutpZ6uNJ90sS2'
+    );
