+++ ex00/src/main/webapp/WEB-INF/application.properties
@@ -1,0 +1,4 @@
+spring.datasource.url=jdbc:postgresql://localhost:5432/fwa
+spring.datasource.username=postgres
+spring.datasource.password=123
+spring.datasource.driver.name=org.postgresql.Driver
