+++ ex00/src/main/webapp/WEB-INF/html/signUp.html
@@ -1,0 +1,87 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8" />
+    <title>FWA</title>
+    <style>
+        form {
+            margin: 200px auto 0;
+            width: 450px;
+            padding: 1em;
+            border: 1px solid #ccc;
+            border-radius: 1em;
+        }
+
+        ul {
+            list-style: none;
+            padding: 0;
+            margin: 0;
+        }
+
+        form li + li {
+            margin-top: 1em;
+        }
+
+        label {
+            display: inline-block;
+            width: 140px;
+            text-align: right;
+        }
+
+        input {
+            font: 1em sans-serif;
+            width: 300px;
+            box-sizing: border-box;
+            border: 1px solid #999;
+        }
+
+        input:focus {
+            border-color: #000;
+        }
+
+        button {
+            padding: 10px 0;
+            border-width: 0;
+            display: block;
+            width: 120px;
+            margin: 25px auto 0;
+            background: #60e6c5;
+            color: white;
+            font-size: 14px;
+            outline: none;
+            text-transform: uppercase;
+        }
+
+    </style>
+</head>
+
+<body>
+<form action="/signUp" method="post">
+    <ul>
+        <li>
+            <label for="name">First name:</label>
+            <input type="text" id="name" name="name" />
+        </li>
+        <li>
+            <label for="surname">Last name:</label>
+            <input type="text" id="surname" name="surname" />
+        </li>
+        <li>
+            <label for="phone">Phone number:</label>
+            <input type="text" id="phone" name="phone" />
+        </li>
+        <li>
+            <label for="mail">E-mail:</label>
+            <input type="email" id="mail" name="email" />
+        </li>
+        <li>
+            <label for="password">Password:</label>
+            <input type="password" id="password" name="password" />
+        </li>
+        <li class="button">
+            <button type="submit">Register</button>
+        </li>
+    </ul>
+</form>
+</body>
+</html>
