+++ ex00/src/main/webapp/WEB-INF/html/welcome.html
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>FWA</title>
+    <style>
+        h2 {
+            padding: 10px 0;
+            border-width: 0;
+            border-radius: 1em;
+            display: block;
+            width: 500px;
+            height: 50px;
+            margin: 200px auto 0;
+            background: #60e6c5;
+            color: black;
+            font-size: 25px;
+            outline: none;
+            text-transform: uppercase;
+            text-align: center;
+
+        }
+    </style>
+</head>
+<body>
+    <h2>You are sign in!</h2>
+</body>
+</html>
